+++ NEWS.md
@@ -1,5 +1,14 @@
 # rlang (development version)
 
+* `abort()`, `warn()`, and `inform()` gain an `.inherit` parameter.
+  This controls whether `parent` is inherited. If `FALSE`,
+  `cnd_inherits()` and `try_fetch()` do not match chained conditions
+  across parents.
+
+  It's normally `TRUE` by default, but if a warning is chained to an
+  error or a message is chained to a warning or error (downgraded
+  chaining), `.inherit` defaults to `FALSE` (#1573).
+
 * `dots_splice()` is deprecated. This function was previously in
   the questioning lifecycle stage as we were moving towards the
   explicit `!!!` splicing style.